FACTOR ANALYSIS  FOR  CATEGORICAL SOILS VARIABLES
This is an try to make continuous categorical soils variables. The idea
is to  run a factor analysis of three types of categorical variables:
soils, geomorphologicals and physicals. And work with the coordinates of
each pixel in the factorial plane

I did a prove  at 10km x10lm pixel, it run well , i tried to do it at
2x2 km but it took more than two weeks in a 32 ram computer. i couldn´t
wait to finish it . It need more proves